requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
groq==0.4.1
redis==5.0.1
python-dotenv==1.0.0